pyproject.toml
[tool.poetry]
name = "smartplug_energy_controller"
version = "0.0.6"
description = "Turning Tapo smartplug on/off depending on current electricity consumption"
authors = ["Heiko Bauer <heiko_bauer@icloud.com>"]
repository = "https://github.com/die-bauerei/smartplug-energy-controller"
readme = "README.md"
packages = [
    {include = "smartplug_energy_controller"},
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.28.0"
plugp100 = "^5.1.3"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
ruamel-yaml = "^0.18.6"

[tool.poetry.group.dev.dependencies]
pydantic = "^2.4.1"
mypy = "^1.5.1"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
